1.6.0: исправления после CI — импорт SliderSetting, float-толщина разметки, replace('_', ' ')

+++ src/main/java/rtx/byazen/api/ui/ScreenshotAnnotatorScreen.java
@@ -64,7 +64,7 @@ extends BaseScreen {
         float x2;
         float y2;
         int color;
-        int thickness;
+        float thickness;
         String text = "";
     }
